# docs/admin/swagger.yaml
basePath: /admin/warehouse
definitions:
...
  requestmodel.SupplierHolidayChangeStatus:
    properties:
      status:
        type: string
    type: object
  requestmodel.SupplierHolidayCreate:
    properties:
      from:
        type: string
      isApplyAll:
        type: boolean
      reason:
        type: string
      supplier:
        type: string
      title:
        type: string
      to:
        type: string
      warehouses:
        items:
          type: string
        type: array
    type: object
  requestmodel.SupplierHolidayUpdate:
    properties:
      from:
        type: string
      isApplyAll:
        type: boolean
      reason:
        type: string
      supplier:
        type: string
      title:
        type: string
      to:
        type: string
      warehouses:
        items:
          type: string
        type: array
    type: object
...
  responsemodel.ResponseChangeStatus:
    properties:
      _id:
        type: string
      status:
        type: string
    type: object
  responsemodel.ResponseCreate:
    properties:
      _id:
        type: string
    type: object
  responsemodel.ResponseInfo:
    properties:
      _id:
        type: string
      name:
        type: string
    type: object
...
  responsemodel.ResponseStaffShort:
    properties:
      _id:
        type: string
      name:
        type: string
    type: object
  responsemodel.ResponseSupplierHolidayAll:
    properties:
      limit:
        type: integer
      list:
        items:
          $ref: '#/definitions/responsemodel.ResponseSupplierHolidayBrief'
        type: array
      total:
        type: integer
    type: object
  responsemodel.ResponseSupplierHolidayBrief:
    properties:
      _id:
        type: string
      createdBy:
        $ref: '#/definitions/responsemodel.ResponseInfo'
      from:
        type: string
      isApplyAll:
        type: boolean
      reason:
        type: string
      source:
        type: string
      status:
        type: string
      supplier:
        $ref: '#/definitions/responsemodel.ResponseSupplierShort'
      title:
        type: string
      to:
        type: string
      warehouses:
        items:
          $ref: '#/definitions/responsemodel.ResponseWarehouseShort'
        type: array
    type: object
  responsemodel.ResponseSupplierHolidayDetail:
    properties:
      _id:
        type: string
      createdBy:
        $ref: '#/definitions/responsemodel.ResponseStaffShort'
      from:
        type: string
      isApplyAll:
        type: boolean
      reason:
        type: string
      source:
        type: string
      status:
        type: string
      supplier:
        $ref: '#/definitions/responsemodel.ResponseSupplierShort'
      title:
        type: string
      to:
        type: string
      warehouses:
        items:
          $ref: '#/definitions/responsemodel.ResponseWarehouseShort'
        type: array
    type: object
  responsemodel.ResponseSupplierShort:
    properties:
      _id:
        type: string
      name:
        type: string
    type: object
  responsemodel.ResponseUpdate:
    properties:
      _id:
        type: string
    type: object
...
  responsemodel.ResponseWarehouseShort:
    properties:
      _id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: dev@selly.vn
    name: Dev team
    url: https://selly.vn
  description: |
    All APIs for Warehouse admin.

    ******************************
    - Add description
    ******************************
  termsOfService: https://selly.vn
  title: Selly Warehouse - Admin API
  version: "1.0"
paths:
...
  /supplier-holidays:
    get:
      consumes:
      - application/json
      operationId: admin-supplier-holiday-all
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      - in: query
        name: fromAt
        type: string
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: string
      - in: query
        name: supplier
        type: string
      - in: query
        name: toAt
        type: string
      - in: query
        name: warehouse
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseSupplierHolidayAll'
      security:
      - ApiKeyAuth: []
      summary: All
      tags:
      - SupplierHoliday
    post:
      consumes:
      - application/json
      operationId: admin-supplier-holiday-create
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SupplierHolidayCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCreate'
      security:
      - ApiKeyAuth: []
      summary: Create
      tags:
      - SupplierHoliday
  /supplier-holidays/{id}:
    get:
      consumes:
      - application/json
      operationId: admin-supplier-holiday-detail
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      - description: Supplier holiday id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseSupplierHolidayDetail'
      security:
      - ApiKeyAuth: []
      summary: Detail
      tags:
      - SupplierHoliday
    put:
      consumes:
      - application/json
      operationId: admin-supplier-holiday-update
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      - description: Supplier holiday id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SupplierHolidayUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseChangeStatus'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - SupplierHoliday
  /supplier-holidays/{id}/status:
    patch:
      consumes:
      - application/json
      operationId: admin-supplier-holiday-change-status
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      - description: Supplier holiday id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SupplierHolidayChangeStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseUpdate'
      security:
      - ApiKeyAuth: []
      summary: Change status
      tags:
      - SupplierHoliday
  /supplier-holidays/run-job-update-holiday-status-for-supplier:
    get:
      consumes:
      - application/json
      operationId: run-job-update-holiday-status-for-supplier
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: RunJobUpdateHolidayStatusForSupplier
      tags:
      - SupplierHoliday
  /supplier-holidays/run-job-update-holiday-warehouses:
    get:
      consumes:
      - application/json
      operationId: run-job-update-holiday-warehouses
      parameters:
      - description: Device id
        in: header
        name: device-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: RunJobUpdateHolidayWarehouses
      tags:
      - SupplierHoliday
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"